refactor orchestrator and stack pkg on aws vpc
