+++ cmd/cmd/corp-microsoft-destroy.go
@@ -4,13 +4,13 @@ import (
 	"github.com/spf13/cobra"
 	"github.com/spf13/viper"
 
-	"github.com/adrianriobo/qenvs/pkg/infra/aws/vpc/stacks"
+	"github.com/adrianriobo/qenvs/pkg/infra/aws/vpc/orchestrator"
 	"github.com/adrianriobo/qenvs/pkg/util/logging"
 )
 
 const (
 	corpMicrosoftDestroyCmdName        string = "destroy"
-	corpMicrosoftCmdDestroyDescription string = "destrioy MS corporate environment"
+	corpMicrosoftCmdDestroyDescription string = "destroy MS corporate environment"
 )
 
 func init() {
@@ -24,7 +24,7 @@ var corpMicrosoftDestroyCmd = &cobra.Command{
 		if err := viper.BindPFlags(cmd.Flags()); err != nil {
 			return err
 		}
-		if err := stacks.DestroyVPC("qenvs", "file:///tmp/qenvs"); err != nil {
+		if err := orchestrator.DestroyNetwork("qenvs", "file:///tmp/qenvs"); err != nil {
 			logging.Error(err)
 		}
 		return nil
