+++ pkg/infra/aws/vpc/stacks/network.go
@@ -1,4 +1,4 @@
-package orchestrator
+package stacks
 
 import (
 	"fmt"
